+++ bindings/pyroot_experimental/cppyy/CPyCppyy/src/TPyArg.cxx
@@ -17,7 +17,7 @@
 void TPyArg::CallConstructor(
     PyObject*& pyself, PyObject* pyclass, const std::vector<TPyArg>& args)
 {
-    int nArgs = args.size();
+    int nArgs = (int)args.size();
     PyObject* pyargs = PyTuple_New(nArgs);
     for (int i = 0; i < nArgs; ++i)
         PyTuple_SET_ITEM(pyargs, i, (PyObject*)args[i]);
@@ -36,7 +36,7 @@ void CallConstructor(PyObject*& pyself, PyObject* pyclass)
 //- generic dispatcher -------------------------------------------------------
 PyObject* TPyArg::CallMethod(PyObject* pymeth, const std::vector<TPyArg>& args)
 {
-    int nArgs = args.size();
+    int nArgs = (int)args.size();
     PyObject* pyargs = PyTuple_New(nArgs);
     for (int i = 0; i < nArgs; ++i)
         PyTuple_SET_ITEM(pyargs, i, (PyObject*)args[i]);
